Used ByteMan in ClientIT (now renamed to ClientResetAndAbortIT) to verify channel.shutdownInput is called rather than trying to verify this by forcing a reset by sending data, which is not reliable across different network stacks. Fixed ReadStream to do defered reset when handling IOException from read to behavior is consistent when reply stream BEGIN gets read after the IOException. Migrated ClientIT and ServerIT off the old Agrona-based scripts.
