+++ pom.xml
@@ -55,7 +55,7 @@
     <nukleus.plugin.version>0.10</nukleus.plugin.version>
     <nukleus.version>0.11</nukleus.version>
 
-    <nukleus.tcp.spec.version>0.16</nukleus.tcp.spec.version>
+    <nukleus.tcp.spec.version>develop-SNAPSHOT</nukleus.tcp.spec.version>
     <reaktor.test.version>0.20</reaktor.test.version>
   </properties>
